Extract deployment session state from the deploy shell
